requirements.txt
psycopg2-binary==2.9.9
fastapi==0.95.0
SQLAlchemy==2.0.15
unicorn==2.0.1
uvicorn==0.22.0
loguru==0.5.3
async-generator==1.10
attrs==23.1.0
blinker==1.6.2
Brotli==1.0.9
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
exceptiongroup==1.1.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.4
kaitaistruct==0.10
outcome==1.2.0
packaging==23.1
pyasn1==0.5.0
pycparser==2.21
pyOpenSSL==23.1.1
pyparsing==3.0.9
PySocks==1.7.1
requests==2.31.0
sniffio==1.3.0
sortedcontainers==2.4.0
tqdm==4.65.0
zstandard==0.21.0
autopep8==2.0.1
colorama==0.4.6
pycodestyle==2.10.0
soupsieve==2.4
tomli==2.0.1
win32-setctime==1.1.0
wsproto==1.2.0
asyncpg==0.28.0
alembic==1.12.1
PyJWT==2.6.0
dnspython==2.4.2 
email-validator==2.1.0.post1
pytest==7.4.3
httpx==0.25.1
python-jose==3.3.0
passlib==1.7.4
pytest-asyncio==0.21.1
python-multipart==0.0.6
python-decouple==3.8